[#1144] Add base OpenSpending theme and methods to overwrite it
To avoid repeating the base theme in every project we need it, I moved it to
this repository. The long vision for this repository is for it to contain all
base stylesheets, images, fonts, and anything not project-specific. Now that
includes the themes.

We also provide methods for a user of `os-styles` to get the themes, also
allowing them to overwrite the base themes' attributes. We do this by merging
the themes in this specific order:

1. osStyles.themes.default
2. osStyles.themes[themeName] (if it exists)
3. userThemes.default (if it exists)
4. userThemes[themeName] (if it exists)

Where `themeName` is the name of the theme the user asked (e.g. theme "OKI").
This allows us to change the theme without having to repeat everything all
other sites need as well. For example, the footer links, social media links,
base colours, etc.

The themes variables can be split in 2 types: content and presentation. Styling
variables would be colours, for example, while content variables are the social
media links, the footer links, etc.

The content variables are easy to implement. The user has them in an object, so
she just needs to add them to the page using her templating engine (Jinja,
Angular, ...). The presentation variables are trickier.

Ideally, I'd like to use the presentation variables in a CSS pre-processor like
SASS, and then overwrite the SASS variables defined in `os-styles`. That isn't
easy, though. There're tools that allow loading a variable from JavaScript into
SASS, but I'd prefer not to require any new tool. This will be implemented in
every different front-end project we develop, so too keep things DRY I want to
require little to no code.

The solution was to follow @akariv's solution. The user must add a
`<style></style>` tag and generate the CSS using the variables from the
templating engine. Hopefully, that'll be in the server, so we won't be adding a
CSS dynamically. If we do, that should result in a terrible flash of misstyled
content. But it'll work.

An important drive for this design was to keep the code DRY. Allowing us to
change the theme's colours in a single place and have it work everywhere.
Another drive was wanting to keep the individual changes in the projects that
use this inside their repository. I didn't want to have a `openspending.json`
theme that had keys for each project, like:

```javascript
{
  "os-explorer": {
    // ...
  },
  "os-viewer": {
    // ...
  },
}
```

Because that means that to change the theme of *os-explorer*, I have to change
the *os-styles* repository, which is quite cumbersome. Keep local variables
in their local repository.

This proposed solution is cumbersome as well, but it's a step in the right
direction IMHO.

openspending/openspending#1144
openspending/openspending#1143